content/canvas/test/test_2d.getcontext.exists.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.getcontext.exists</title>
<!-- Testing: The 2D context is implemented -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(canvas.getContext('2d') !== null, "canvas.getContext('2d') !== null");

SimpleTest.finish();

});
</script>
